feat(docs): more doctests, docstrings
