+++ common/src/ether/compiler.rs
@@ -99,3 +99,85 @@ pub fn detect_compiler(bytecode: &str) -> (&'static str, String) {
 
     (compiler, version.trim_end_matches('.').to_string())
 }
+
+#[cfg(test)]
+mod test_compiler {
+    use crate::ether::compiler::detect_compiler;
+
+    #[test]
+    fn test_detect_compiler_proxy_minimal() {
+        let bytecode = "363d3d373d3d3d363d73";
+        let expected_result = ("proxy", "minimal".to_string());
+        assert_eq!(detect_compiler(bytecode), expected_result);
+    }
+
+    #[test]
+    fn test_detect_compiler_proxy_vyper() {
+        let bytecode = "366000600037611000600036600073";
+        let expected_result = ("proxy", "vyper".to_string());
+        assert_eq!(detect_compiler(bytecode), expected_result);
+    }
+
+    #[test]
+    fn test_detect_compiler_vyper_range_1() {
+        let bytecode = "6004361015";
+        let expected_result = ("vyper", "0.2.0-0.2.4,0.2.11-0.3.3".to_string());
+        assert_eq!(detect_compiler(bytecode), expected_result);
+    }
+
+    #[test]
+    fn test_detect_compiler_vyper_range_2() {
+        let bytecode = "341561000a";
+        let expected_result = ("vyper", "0.2.5-0.2.8".to_string());
+        assert_eq!(detect_compiler(bytecode), expected_result);
+    }
+
+    #[test]
+    fn test_detect_compiler_solc_range_1() {
+        let bytecode = "731bf797";
+        let expected_result = ("solc", "0.4.10-0.4.24".to_string());
+        assert_eq!(detect_compiler(bytecode), expected_result);
+    }
+
+    #[test]
+    fn test_detect_compiler_solc_range_2() {
+        let bytecode = "6080604052";
+        let expected_result = ("solc", "0.4.22+".to_string());
+        assert_eq!(detect_compiler(bytecode), expected_result);
+    }
+
+    #[test]
+    fn test_detect_compiler_solc_range_3() {
+        let bytecode = "6060604052";
+        let expected_result = ("solc", "0.4.11-0.4.21".to_string());
+        assert_eq!(detect_compiler(bytecode), expected_result);
+    }
+
+    #[test]
+    fn test_detect_compiler_vyper() {
+        let bytecode = r#"7679706572"#;
+        let expected_result = ("vyper", "unknown".to_string());
+        assert_eq!(detect_compiler(bytecode), expected_result);
+    }
+
+    #[test]
+    fn test_detect_compiler_solc() {
+        let bytecode = "736f6c63";
+        let expected_result = ("solc", "unknown".to_string());
+        assert_eq!(detect_compiler(bytecode), expected_result);
+    }
+
+    #[test]
+    fn test_detect_compiler_solc_metadata() {
+        let bytecode = "736f6c63434d4e";
+        let expected_result = ("solc", "unknown".to_string());
+        assert_eq!(detect_compiler(bytecode), expected_result);
+    }
+
+    #[test]
+    fn test_detect_compiler_vyper_metadata() {
+        let bytecode = "7679706572833135353030";
+        let expected_result = ("vyper", "49.53.53".to_string());
+        assert_eq!(detect_compiler(bytecode), expected_result);
+    }
+}
