+++ common/src/ether/evm/core/mod.rs
@@ -3,6 +3,5 @@ pub mod memory;
 pub mod opcodes;
 pub mod stack;
 pub mod storage;
-mod tests;
 pub mod types;
 pub mod vm;
