+++ common/src/ether/mod.rs
@@ -4,4 +4,3 @@ pub mod lexers;
 pub mod rpc;
 pub mod selectors;
 pub mod signatures;
-pub mod tests;
